+++ src/apps/toolbar/modules/Workspaces/index.tsx
@@ -1,9 +1,9 @@
 import { WorkspacesTM, WorkspaceTMMode } from '../../../shared/schemas/Placeholders';
 import { cx } from '../../../shared/styles';
 import { invoke } from '@tauri-apps/api/core';
+import { Tooltip } from 'antd';
+import { Reorder } from 'framer-motion';
 import { useSelector } from 'react-redux';
-
-import { Item } from '../item/infra';
 
 import { Selectors } from '../shared/store/app';
 
@@ -19,7 +19,7 @@ export function WorkspacesModule({ module }: Props) {
 
   if (mode === WorkspaceTMMode.Dotted) {
     return (
-      <div className="ft-bar-item">
+      <Reorder.Item as="div" value={module} className="ft-bar-item" style={module.style}>
         <ul className="ft-bar-item-content workspaces">
           {workspaces.map((_, idx) => (
             <li
@@ -31,21 +31,36 @@ export function WorkspacesModule({ module }: Props) {
             />
           ))}
         </ul>
-      </div>
+      </Reorder.Item>
     );
   }
 
-  return workspaces.map((name, idx) => {
-    return (
-      <Item
-        key={name}
-        active={idx === activeWorkspace}
-        module={{
-          ...module,
-          onClick: `switch-workspace -> ${idx}`,
-          template: mode === WorkspaceTMMode.Named ? `"${name}"` : `"${idx + 1}"`,
-        }}
-      />
-    );
-  });
+  return (
+    <Reorder.Item as="div" id={module.id} value={module}>
+      {workspaces.map((name, idx) => {
+        return (
+          <Tooltip
+            arrow={false}
+            mouseLeaveDelay={0}
+            overlayClassName="ft-bar-item-tooltip"
+            title={name}
+            key={name}
+          >
+            <div
+              style={module.style}
+              className={cx('ft-bar-item', {
+                'ft-bar-item-clickable': true,
+                'ft-bar-item-active': idx === activeWorkspace,
+              })}
+              onClick={() => invoke('switch_workspace', { idx })}
+            >
+              <div className="ft-bar-item-content">
+                {mode === WorkspaceTMMode.Named ? `${name}` : `${idx + 1}`}
+              </div>
+            </div>
+          </Tooltip>
+        );
+      })}
+    </Reorder.Item>
+  );
 }